+++ sdk/@launchdarkly/observability-python/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the LaunchDarkly Python Observability will be documented in this file. This project adheres to Semantic Versioning.
 
+## [0.1.1](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-python-0.1.0...launchdarkly-observability-python-0.1.1) (2025-08-28)
+
+
+### Bug Fixes
+
+* Create span when no active span for error. ([#191](https://github.com/launchdarkly/observability-sdk/issues/191)) ([0341b0f](https://github.com/launchdarkly/observability-sdk/commit/0341b0f62ea1f88ab286ad14f9d72bc0b297e668))
+
 ## 0.1.0 (2025-07-18)